feat(esl-toggleable): introducing a base trigger class in purpose to simplify and make Trigger-Toggleable constructions more API strict in ESL
